Refactor "impl" to "player"
