+++ src/main/java/xyz/srgnis/bodyhealthsystem/body/player/PlayerBody.java
@@ -1,9 +1,8 @@
-package xyz.srgnis.bodyhealthsystem.body.impl;
+package xyz.srgnis.bodyhealthsystem.body.player;
 
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.Identifier;
 import xyz.srgnis.bodyhealthsystem.body.Body;
-import xyz.srgnis.bodyhealthsystem.body.impl.parts.*;
+import xyz.srgnis.bodyhealthsystem.body.player.parts.*;
 
 public class PlayerBody extends Body {
