+++ src/main/java/xyz/srgnis/bodyhealthsystem/body/player/parts/LegBodyPart.java
@@ -1,27 +1,27 @@
-package xyz.srgnis.bodyhealthsystem.body.impl.parts;
+package xyz.srgnis.bodyhealthsystem.body.player.parts;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.util.Identifier;
 import xyz.srgnis.bodyhealthsystem.body.BodyPart;
-import xyz.srgnis.bodyhealthsystem.body.impl.PlayerBodyParts;
+import xyz.srgnis.bodyhealthsystem.body.player.PlayerBodyParts;
 
-public class FootBodyPart extends BodyPart {
+public class LegBodyPart extends BodyPart {
     private static float defaultMaxHealth = 4;
 
     //TODO: think other way to do this
     public static Identifier getSide(String side){
         if(side == "left") {
-            return PlayerBodyParts.LEFT_FOOT;
+            return PlayerBodyParts.LEFT_LEG;
         }else{
-            return PlayerBodyParts.RIGHT_FOOT;
+            return PlayerBodyParts.RIGHT_LEG;
         }
     }
 
-    public FootBodyPart(String side, Entity entity) {
+    public LegBodyPart(String side, Entity entity) {
         super(defaultMaxHealth, defaultMaxHealth, entity, getSide(side));
     }
 
-    public FootBodyPart(String side, float maxHealth, float health, Entity entity) {
+    public LegBodyPart(String side, float maxHealth, float health, Entity entity) {
         super(maxHealth, health, entity,  getSide(side));
     }
 }
